# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Product Inventory API
  description: |
    A comprehensive hybrid ORM API for product inventory management using ASP.NET Core, 
    Entity Framework Core, and SQL Server stored procedures.
    
    ## Features
    - **Hybrid ORM Architecture**: Combines Entity Framework Core with stored procedures
    - **JWT Authentication**: Secure token-based authentication
    - **Role-based Authorization**: Fine-grained access control
    - **Analytics & Reporting**: Advanced analytics using stored procedures
    - **Inventory Management**: Complete inventory tracking and movement recording
    - **Health Checks**: Built-in monitoring and health endpoints
    
    ## Architecture
    The API implements Clean Architecture principles with:
    - **Entity Framework Core**: Standard CRUD operations and relationships
    - **Stored Procedures**: Complex analytics and reporting
    - **Raw SQL**: Performance-critical operations
    - **Repository Pattern**: Abstracted data access layer
    
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@productinventory.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.productinventory.com/api/v1
    description: Production server
  - url: https://staging-api.productinventory.com/api/v1
    description: Staging server
  - url: https://localhost:5001/api/v1
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /analytics/categories/{id}/statistics:
    get:
      tags:
        - Analytics
      summary: Get category statistics by ID
      description: Get statistics for a specific category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Category statistics retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CategoryStatistics'
...
  /analytics/categories/top-by-product-count:
    get:
      tags:
        - Analytics
      summary: Get top categories by product count
      description: Get top categories ranked by number of products
      parameters:
        - name: limit
          in: query
          description: Number of top categories to return
          schema:
            type: integer
            format: int32
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Top categories retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TopCategory'
  /analytics/categories/bulk-update-status:
    put:
      tags:
        - Analytics
      summary: Bulk update category status
      description: Update the active status of multiple categories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkUpdateCategoryStatusRequest'
      responses:
        '200':
          description: Categories updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/BulkUpdateResult'

  # Inventory Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from the login endpoint
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Optional success/error message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        correlationId:
          type: string
          format: uuid
          description: Request correlation ID for tracking
      required:
        - success
        - timestamp
        - correlationId
...
    CategoryStatistics:
      type: object
      properties:
        categoryId:
          type: integer
          format: int32
          description: Category ID
        categoryName:
          type: string
          description: Category name
        productCount:
          type: integer
          format: int32
          description: Total number of products
        activeProductCount:
          type: integer
          format: int32
          description: Number of active products
        totalInventoryValue:
          type: number
          format: decimal
          description: Total inventory value
        averageProductPrice:
          type: number
          format: decimal
          description: Average product price
        lowStockProductCount:
          type: integer
          format: int32
          description: Number of products with low stock
        outOfStockProductCount:
          type: integer
          format: int32
          description: Number of out-of-stock products
        lastUpdated:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - categoryId
        - categoryName
        - productCount
        - activeProductCount
        - totalInventoryValue
        - averageProductPrice
        - lowStockProductCount
        - outOfStockProductCount
        - lastUpdated
...
    TopCategory:
      type: object
      properties:
        categoryId:
          type: integer
          format: int32
          description: Category ID
        categoryName:
          type: string
          description: Category name
        productCount:
          type: integer
          format: int32
          description: Number of products
        rank:
          type: integer
          format: int32
          description: Ranking position
      required:
        - categoryId
        - categoryName
        - productCount
        - rank
    BulkUpdateCategoryStatusRequest:
      type: object
      properties:
        categoryIds:
          type: array
          items:
            type: integer
            format: int32
          description: List of category IDs to update
        isActive:
          type: boolean
          description: New active status
      required:
        - categoryIds
        - isActive
    BulkUpdateResult:
      type: object
      properties:
        updatedCount:
          type: integer
          format: int32
          description: Number of categories updated
        failedCount:
          type: integer
          format: int32
          description: Number of updates that failed
        errors:
          type: array
          items:
            type: string
          description: List of error messages for failed updates
      required:
        - updatedCount
        - failedCount

    # Inventory Summary Schemas
...